Moved entity and response classes form InterfaceWithDatabase class into separate folders. Added swagger ui for smoke testing.
