+++ AcmeStudios.ApiRefactor/DTOs/UpdateStudioItemDto.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.ComponentModel.DataAnnotations;
+using AcmeStudios.ApiRefactor.Entities;
 
 namespace AcemStudios.ApiRefactor.DTOs
 {
