+++ AcmeStudios.ApiRefactor/Entities/StudioItem.cs
@@ -1,26 +1,29 @@
 ﻿using System;
 using System.ComponentModel.DataAnnotations;
-using AcmeStudios.ApiRefactor.Entities;
+using System.ComponentModel.DataAnnotations.Schema;
 
-namespace AcemStudios.ApiRefactor.DTOs
+namespace AcmeStudios.ApiRefactor.Entities
 {
-    public class GetStudioItemDto
+    public class StudioItem
     {
+        [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public int StudioItemId { get; set; }
         public DateTime Acquired { get; set; }
-        public DateTime? Sold { get; set; }
+        public DateTime? Sold { get; set; } = null;
         [Required]
         public string Name { get; set; }
         [Required]
         public string Description { get; set; }
         [Required]
         public string SerialNumber { get; set; }
-        public decimal Price { get; set; }
-        public decimal? SoldFor { get; set; }
-        public bool Eurorack { get; set; }
+        public decimal Price { get; set; } //= 10.00M;
+        public decimal? SoldFor { get; set; } //= 0M;
+        public bool Eurorack { get; set; } //= false;
+        [Required]
         public int StudioItemTypeId { get; set; }
         public StudioItemType StudioItemType { get; set; }
 
-        //public StudioItemImage StudioItemImage { get; set; }
+
     }
+
 }
